lint, make models more modular
