+++ alembic/env.py
@@ -6,7 +6,7 @@ from sqlalchemy.engine import Connection
 from sqlalchemy.ext.asyncio import async_engine_from_config
 
 from alembic import context
-from obearon.database.models import Base
+from obearon.database.engine import Base
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
