+++ alembic/versions/ccef2015d4be_create_initial_tables.py
@@ -8,9 +8,9 @@ Create Date: 2025-04-20 16:19:58.182605
 
 from typing import Sequence, Union
 
-from alembic import op
 import sqlalchemy as sa
 
+from alembic import op
 
 # revision identifiers, used by Alembic.
 revision: str = "ccef2015d4be"
